docs `installation.md`: describe how to pass options to the installer on Linux
I'd feel more comfortable using the installer with the instructions in this commit, since I'm uncomfortable with random scripts trying to modify my system config (PATH in this case).

Currently, the installer seems to be the best way to install `uv` that allows updating it on a system without Homebrew or `pipx`. I hope somebody will provide similar instructions for Windows.

I considered recommending saving the script to a file and then running that, but I think it's better to have fewer options in the instructions. Most people who'd want to save the file would figure it out.

As an aside, I would personally appreciate if `uv` could be installed easily with `cargo install` or `cargo binstall`, but a friendly script that acts predictably is probably more useful for more people.