Extract METADATA reading into a crate
This is preparatory work for the upload functionality, which needs to read the METADATA file and attach its parsed contents to the POST request: We move finding the `.dist-info` from `install-wheel-rs` and `uv-client` to a new `uv-metadata` crate, so it can be shared with the publish crate.

This isn't properly ready since the upload code isn't ready, but i'm PR-ing it now because it already had merge conflicts.