Don't show deploy notification when we don't need to
To scope permissions more tightly, we run the upload test in an environment. This makes GitHub show `{user} deployed to uv-test-publish` messages, even though we skipped the actual publish step. We now guard the whole action, so the pseudo-deploy is only shown when the publishing code changed and we run that trusted publishing test.