Check consistency between prepare and build step
PEP 517 mandates that the metadata must be consistent between `prepare_metadata_for_build_wheel` and `build_wheel` by passing the directory written in the prepare step to the build step. There is no reason why we would violate this guarantee, but we are prudent and check that `METADATA` and `entry_points.txt` actually match.