fix: Improve compatibility with VSCode PS1 prompt
This was previously reported in #3037, and the repro there still works
-- create any venv using uv and the prompt will be `nameoriginalps1`
instead of `(name) originalps1`. VSCode fixed one of the two bugs
reported there.

The VSCode-generated activate script is pretty different from the
uv-generated activate script. The most important thing seems to be that
vscode's prompt takes the VIRTUAL_ENV_PROMPT directly instead of obeying
PS1, from VSCodes activate script:

```shell
if [ -z "${VIRTUAL_ENV_DISABLE_PROMPT:-}" ] ; then
    _OLD_VIRTUAL_PS1="${PS1:-}"
    PS1="(.venv) ${PS1:-}"
    export PS1
    VIRTUAL_ENV_PROMPT="(.venv) "
    export VIRTUAL_ENV_PROMPT
fi
```

This changes uv to use VIRTUAL_ENV_PROMPT directly, which does fix
vscode for me.

This changes uv to be different from [what pypa/virtualenv does][1], so
I'm not really sure if it should be accepted or if a bug should be
reported to VSCode.

[1]: https://github.com/pypa/virtualenv/blob/b56d09248aecf584e985956185afb6745a7daac7/src/virtualenv/activation/bash/activate.sh#L71-L75