Enable HTTP/2 in reqwest
We should support HTTP/2, a default feature of reqwest, which as it seems was disabled as an oversight. With our many and parallel requests, this should give us even a smaller perf improvement, or at least decrease the load on the server/CDN by multiplexing and smaller headers (see e.g. https://www.cloudflare.com/learning/performance/http2-vs-http1.1/)