Use `dev-dependencies` and `requires-dev` for lockfile compatibility (#8501)
## Summary

Per discussion on Discord, we're going to keep these names for now
(changed in #8391 but not yet released) for compatibility. They're just
cosmetic, but with the changes as-is, if you ran an older uv version
over a newer lockfile...

- For `uv sync`: the lockfile would be invalidated, and we'd re-resolve.
- For `uv sync --frozen`: we'd silently skip the development
dependencies.

We'll change these names in the future once we've added better error
handling around lockfile versions (#8465).