Start using the version ranges crate
See https://github.com/pubgrub-rs/pubgrub/pull/262 for prior discussion.

In most parts of uv, we don't want to depend on pubgrub, but we want to use its powerful version ranges arithmetic. https://github.com/pubgrub-rs/pubgrub/pull/262 split out the ranges into a separate, small crate (that can be published separately to crates.io).

With this first in a series of changes, we only depend on pubgrub in the uv-resolver crate.

Note that the name is now `Ranges`, since the type formerly known as `Range` was a misnormer, it contains multiple ranges.