uv-resolver: add filtering step after forks are combined
While we do filter based on Python markers during forking, it turns out
that the markers for each fork are "combined" *after* this filtering
step. But the process of combination can result in a more specific
marker that is always false for the configured Python requirement. This
could result in dependencies with markers that are always false (like
`python_version < '0'`) appearing in the resolution.

I think it would probably be better to do this check during forking, but
that's a bit of a bigger refactor.